+++ Source/Modules/ISHServer/Get-ISHPrerequisites.ps1
@@ -113,6 +113,7 @@ function Get-ISHPrerequisites
         if($PSCmdlet.MyInvocation.MyCommand.Module.Name -eq "ISHServer.14")
         {
             $filesToDownload+=Get-Variable -Name "ISHServer:AdoptOpenJDK" -ValueOnly
+            $filesToDownload+=Get-Variable -Name "ISHServer:AdoptOpenJRE" -ValueOnly
         }
 
         $osInfo=Get-ISHOSInfo