+++ Source/Modules/ISHServer/Get-ISHPrerequisites.ps1
@@ -121,6 +121,7 @@ function Get-ISHPrerequisites
         {
             $filesToDownload+=Get-Variable -Name "ISHServer:AdoptOpenJDK" -ValueOnly
             $filesToDownload+=Get-Variable -Name "ISHServer:AdoptOpenJRE" -ValueOnly
+            $filesToDownload+=Get-Variable -Name "ISHServer:MSOLEDBSQL" -ValueOnly
         }
 
         #Dependend on Operating System Information (OS Server vesion, already installed prerequisites)