+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 **1.7**
 
 Issues:
-- GH-48: SAdd installaton of AdoptOpenJDK pre-requisite
+- GH-48: SAdd installaton of AdoptOpenJDK and JRE pre-requisite
 
 **1.6**