+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 **1.7**
 
 Issues:
-- GH-48: SAdd installaton of AdoptOpenJDK and JRE pre-requisite
+- GH-48: Add installaton of AdoptOpenJDK and JRE pre-requisite
+- GH-50: Upgrade .NET Framework pre-requisite to 4.7.2 for the upcoming release of Tridion Docs 14.0.0
 
 **1.6**